+++ R/spread_values.R
@@ -105,7 +105,8 @@ json_factory <- function(map.function) {
 #' Navigates nested objects to get at names of a specific type, to be used as
 #' arguments to \code{\link{spread_values}}
 #'
-#' Note that these functions fail if they encounter the incorrect type.
+#' Note that these functions fail if they encounter the incorrect type. Note
+#' that \code{jnumber()} is an alias for \code{jdouble()}.
 #'
 #' @seealso \code{\link{spread_values}} for using these functions to spread
 #'          the values of a JSON object into new columns
@@ -120,12 +121,21 @@ NULL
 
 #' @rdname json_functions
 #' @export
-jstring <- json_factory(map_chr)
+jstring <- json_factory(purrr::map_chr)
 
 #' @rdname json_functions
 #' @export
-jnumber <- json_factory(map_dbl)
+jlogical <- json_factory(purrr::map_lgl)
 
 #' @rdname json_functions
 #' @export
-jlogical <- json_factory(map_lgl)
+jinteger <- json_factory(purrr::map_int)
+
+#' @rdname json_functions
+#' @export
+jdouble <- json_factory(purrr::map_dbl)
+
+#' @rdname json_functions
+#' @export
+jnumber <- jdouble
+